+++ simpleclient-1.19.4/src/main/java/simpleclient/mixin/feature/legacypvp/GuiMixin.java
@@ -1,9 +1,9 @@
 package simpleclient.mixin.feature.legacypvp;
 
+import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.AttackIndicatorStatus;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Gui;
-import net.minecraft.client.gui.GuiGraphics;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -20,7 +20,7 @@ abstract class GuiMixin {
     private AttackIndicatorStatus attackIndicator;
 
     @Inject(at = @At(value = "HEAD"), method = "render")
-    public void render1(GuiGraphics guiGraphics, float patrialTicks, CallbackInfo ci) {
+    public void render1(PoseStack poseStack, float patrialTicks, CallbackInfo ci) {
         if (LegacyPvP.ENABLED) {
             attackIndicator = minecraft.options.attackIndicator().get();
             minecraft.options.attackIndicator().set(AttackIndicatorStatus.OFF);
@@ -28,7 +28,7 @@ abstract class GuiMixin {
     }
 
     @Inject(at = @At(value = "TAIL"), method = "render")
-    public void render2(GuiGraphics guiGraphics, float partialTicks, CallbackInfo ci) {
+    public void render2(PoseStack poseStack, float partialTicks, CallbackInfo ci) {
         if (LegacyPvP.ENABLED) minecraft.options.attackIndicator().set(attackIndicator);
     }
 }