+++ changelog.md
@@ -1,2 +1,3 @@
 ## SimpleClient %version%
-- Fixed server IP in Discord integration
+- Made server IP in Discord integration lower case
+- Added LegacyPvP for >=1.19.4<br>Can be enabled by the server